+++ samples/vuejs/typescript/00 Custom Validator/src/common/validations/arrayValidation.ts
@@ -1,9 +1,10 @@
 import {FieldValidationResult} from 'lc-form-validation';
 
 export const hasItems = (message) => (value: any[]): FieldValidationResult => {
+  const isValid = value.length > 0;
   return {
     type: 'ARRAY_HAS_ITEMS',
-    succeeded: value.length > 0,
-    errorMessage: message,
+    succeeded: isValid,
+    errorMessage: isValid ? '' : message,
   }
 }
