+++ samples/vuejs/typescript/00 Custom Validator/src/pages/recipe/edit/components/index.ts
@@ -1,0 +1,5 @@
+import {FormComponent} from './form';
+
+export {
+  FormComponent,
+}
