+++ samples/vuejs/typescript/00 Custom Validator/src/pages/recipe/edit/components/ingredientList.tsx
