+++ samples/vuejs/typescript/00 Custom Validator/src/pages/recipe/edit/page.tsx
@@ -1,5 +1,5 @@
 import Vue from 'vue';
-import {FormComponent} from './form';
+import {FormComponent} from './components';
 
 export const EditRecipePage = Vue.extend({
   props: [
