+++ samples/vuejs/typescript/00 Custom Validator/src/pages/recipe/edit/validations/editFormValidation.ts
@@ -1,7 +1,7 @@
 import {
   ValidationConstraints, createFormValidation, Validators
 } from 'lc-form-validation';
-import {hasItems} from './arrayValidation';
+import {hasItems} from '../../../../common/validations/arrayValidation';
 
 const constraints: ValidationConstraints = {
   fields: {
